Move parsed_json_writer to its own file
